[FIX] event: allow event user to confirm registrations even with mail schedulers
Due to ACLs it is currently impossible for event users to confirm a draft
registration if schedulers are involved. Indeed it implies updating the
scheduler which is not do-able for them.

In this commit we sudo the update of schedulers, as this is a technical
object and access is granted through registrations.

Related to Task ID-2414658
COM PR odoo/odoo#68158

X-original-commit: 73d09268922a7d242af85e36a3d81ed25ab4ca33